Major update

+++ deployment/powershell/AddDBRPImage.ps1
@@ -25,16 +25,23 @@ param (
     [String] $databaseName,
 
     [Parameter(Mandatory = $true)]
-    [String] $tableName
+    [String] $tableName,
+
+    [parameter(Mandatory = $false)]
+    [String] $skipMySQL,
+
+    [parameter(Mandatory = $false)]
+    [String] $skipMSSQL
 )
 
 $Global:VerbosePreference = "Continue"
 $Global:ErrorActionPreference = 'Stop'
 $Global:ProgressPreference = 'SilentlyContinue'
 
-$logFolder = "AddVMExtensions"
+$logFolder = "AddDBRPImage"
 $logName = $logFolder
 $progressName = $logFolder
+$skipAddOnDownload = $false
 
 ### SET LOG LOCATION ###
 $logDate = Get-Date -Format FileDate
@@ -54,13 +61,22 @@ Write-Host "Log started at $runTime"
 $progressStage = $progressName
 $progressCheck = CheckProgress -progressStage $progressStage
 
-if (($registerAzS -eq $true) -and ($deploymentMode -ne "Offline")) {
-    if (($progressCheck -eq "Incomplete") -or ($progressCheck -eq "Failed")) {
+if ((($skipMySQL -eq $true) -and ($skipMSSQL -eq $true)) -or ($registerAzS -ne $true)) {
+    $skipAddOnDownload = $true
+}
+
+if (($registerAzS -eq $true) -and ($deploymentMode -ne "Offline") -and ($skipAddOnDownload -ne $true)) {
+    if ($progressCheck -ne "Complete") {
         try {
             if ($progressCheck -eq "Failed") {
                 StageReset -progressStage $progressStage
             }
-
+            elseif ($progressCheck -eq "Skipped") {
+                Write-Host "Operator previously skipped this step, but now wants to perform this step. Updating AzSPoC database to Incomplete."
+                # Update the AzSPoC database back to incomplete
+                StageReset -progressStage $progressStage
+                $progressCheck = CheckProgress -progressStage $progressStage
+            }
             Write-Host "Clearing previous Azure/Azure Stack logins"
             Get-AzContext -ListAvailable | Where-Object { $_.Environment -like "Azure*" } | Remove-AzContext -Force | Out-Null
             Clear-AzContext -Scope CurrentUser -Force
@@ -99,29 +115,35 @@ if (($registerAzS -eq $true) -and ($deploymentMode -ne "Offline")) {
             $activationRG = "azurestack-activation"
             Write-Host "Checking if Azure Stack is activated and successfully registered"
             if ($(Get-AzsAzureBridgeActivation -Name $activationName -ResourceGroupName $activationRG -ErrorAction SilentlyContinue -Verbose)) {
-                Write-Host "Adding Microsoft VM Extensions from the from the Azure Stack Marketplace"
-                $getExtensions = ((Get-AzsAzureBridgeProduct -ActivationName $activationName -ResourceGroupName $activationRG -ErrorAction SilentlyContinue -Verbose | Where-Object { ($_.ProductKind -eq "virtualMachineExtension") -and ($_.Name -like "*microsoft*") }).Name) -replace "default/", ""
-                foreach ($extension in $getExtensions) {
-                    while (!$(Get-AzsAzureBridgeDownloadedProduct -Name $extension -ActivationName $activationName -ResourceGroupName $activationRG -ErrorAction SilentlyContinue -Verbose)) {
-                        Write-Host "Didn't find $extension in your gallery. Downloading from the Azure Stack Marketplace"
-                        Invoke-AzsAzureBridgeProductDownload -ActivationName $activationName -Name $extension -ResourceGroupName $activationRG -Confirm:$false -Verbose
-                        Connect-AzAccount -Environment "AzureStackAdmin" -Tenant $tenantID -Credential $azsCreds -ErrorAction Stop | Out-Null
-                        $sub = Get-AzSubscription | Where-Object { $_.Name -eq "Default Provider Subscription" }
-                        Get-AzSubscription -SubscriptionID $sub.SubscriptionId | Set-AzContext
+                Write-Host "Adding Microsoft AzureStack Add-on RP Windows Server image the Azure Stack Marketplace"
+                $getImage = ((Get-AzsAzureBridgeProduct -ActivationName $activationName -ResourceGroupName $activationRG -ErrorAction SilentlyContinue -Verbose | `
+                            Where-Object { ($_.ProductKind -eq "virtualMachine") -and ($_.Name -like "*microsoft*addonrp*") }).Name) -replace "default/", "" | `
+                    Select-Object -Last 1
+                while (!$(Get-AzsAzureBridgeDownloadedProduct -Name $getImage -ActivationName $activationName -ResourceGroupName $activationRG -ErrorAction SilentlyContinue -Verbose).ProvisioningState -eq 'Succeeded') {
+                    if ($(Get-AzsAzureBridgeDownloadedProduct -Name $getImage -ActivationName $activationName -ResourceGroupName $activationRG -ErrorAction SilentlyContinue -Verbose).ProvisioningState -eq 'Failed') {
+                        Write-Host "There appears to be an unsucessful download of $getImage in your gallery. Cleaning up."
+                        Remove-AzsAzureBridgeDownloadedProduct -Name $getImage -ActivationName $activationName -ResourceGroupName $activationRG -ErrorAction SilentlyContinue -Verbose
                     }
+                    elseif ($(Get-AzsAzureBridgeDownloadedProduct -Name $getImage -ActivationName $activationName -ResourceGroupName $activationRG -ErrorAction SilentlyContinue -Verbose).ProvisioningState -eq 'Canceled') {
+                        Write-Host "There appears to be an unsucessful download of $getImage in your gallery. Cleaning up."
+                        Remove-AzsAzureBridgeDownloadedProduct -Name $getImage -ActivationName $activationName -ResourceGroupName $activationRG -ErrorAction SilentlyContinue -Verbose
+                    }
+                    Write-Host "There's no $getImage image in your gallery. Downloading from the Azure Stack Marketplace"
+                    Invoke-AzsAzureBridgeProductDownload -ActivationName $activationName -Name $getImage -ResourceGroupName $activationRG -Confirm:$false -Verbose
+                    Connect-AzAccount -Environment "AzureStackAdmin" -Tenant $tenantID -Credential $azsCreds -ErrorAction Stop | Out-Null
+                    $sub = Get-AzSubscription | Where-Object { $_.Name -eq "Default Provider Subscription" }
+                    Get-AzSubscription -SubscriptionID $sub.SubscriptionId | Set-AzContext
                 }
-                $getDownloads = (Get-AzsAzureBridgeDownloadedProduct -ActivationName $activationName -ResourceGroupName $activationRG -ErrorAction SilentlyContinue -Verbose | Where-Object { ($_.ProductKind -eq "virtualMachineExtension") -and ($_.Name -like "*microsoft*") })
-                Write-Host "Your Azure Stack gallery now has the following Microsoft VM Extensions for enhancing your deployments:`r`n"
-                foreach ($download in $getDownloads) {
-                    Write-Host "$($download.DisplayName) | Version: $($download.ProductProperties.Version)"
-                }
+                $confirmDownload = (Get-AzsAzureBridgeDownloadedProduct -ActivationName $activationName -ResourceGroupName $activationRG -ErrorAction SilentlyContinue -Verbose | Where-Object { ($_.ProductKind -eq "virtualMachine") -and ($_.Name -like "*microsoft*addonrp*") })
+                Write-Host "Your Azure Stack gallery now has the following Microsoft AzureStack Add-on RP Windows Server image for use with the database RPs:`r`n"
+                Write-Host "$($confirmDownload.DisplayName) | Version: $($confirmDownload.OfferVersion)"
                 # Update the AzSPoC database with successful completion
                 StageComplete -progressStage $progressStage
             }
             else {
                 # No Azure Bridge Activation Record found - Skip rather than fail
-                Write-Host "Skipping Microsoft VM Extension download, no Azure Bridge Activation Object called $activationName could be found within the resource group $activationRG on your Azure Stack"
-                Write-Host "Assuming registration of this Azure Stack POC was successful, you should be able to manually download the VM extensions from Marketplace Management in the admin portal`r`n"
+                Write-Host "Skipping Microsoft AzureStack Add-on RP Windows Server image download, no Azure Bridge Activation Object called $activationName could be found within the resource group $activationRG on your Azure Stack"
+                Write-Host "Assuming registration of this Azure Stack POC was successful, you should be able to manually download the Microsoft AzureStack Add-on RP Windows Server image from Marketplace Management in the admin portal, then rerun the script`r`n"
                 # Update the AzSPoC database with skip status
                 StageSkipped -progressStage $progressStage
             }
@@ -133,15 +155,23 @@ if (($registerAzS -eq $true) -and ($deploymentMode -ne "Offline")) {
             return
         }
     }
-    elseif ($progressCheck -eq "Skipped") {
-        Write-Host "Azure Stack POC Configurator Stage: $progressStage previously skipped"
-    }
-    elseif ($progressCheck -eq "Complete") {
-        Write-Host "Azure Stack POC Configurator Stage: $progressStage previously completed successfully"
+    elseif (($skipAddOnDownload) -and ($progressCheck -ne "Complete")) {
+        Write-Host "Operator chose to skip Resource Provider Deployment"
+        # Update the AzSPoC database with skip status
+        $progressStage = $progressName
+        StageSkipped -progressStage $progressStage
     }
 }
-elseif ($registerAzS -eq $false) {
-    Write-Host "Skipping VM Extension download, as Azure Stack has not been registered`r`n"
+else {
+    if ($registerAzS -eq $false) {
+        Write-Host "Skipping Microsoft AzureStack Add-on RP Windows Server image download, as Azure Stack has not been registered`r`n"
+    }
+    elseif ($deploymentMode -eq "Offline") {
+        Write-Host "Skipping Microsoft AzureStack Add-on RP Windows Server image download, as this is an offline deployment`r`n"
+    }
+    elseif ($skipAddOnDownload -eq $true) {
+        Write-Host "Skipping Microsoft AzureStack Add-on RP Windows Server image download, operator is skipping both SQL and MySQL deployments`r`n"
+    }
     # Update the AzSPoC database with skip status
     StageSkipped -progressStage $progressStage
 }
